robot_interaction: split handler into own file
